set up typescript project
