--- frontend/src/auth/context/AuthProvider.jsx
@@ -1,13 +1,0 @@
-import { createContext, useState } from "react";
-
-const AuthContext = createContext({});
-
-// eslint-disable-next-line react/prop-types
-export function AuthProvider({ children }) {
-  const [auth, setAuth] = useState({});
-  return (
-    <AuthContext.Provider value={{ auth, setAuth }}>{children}</AuthContext.Provider>
-  );
-}
-
-export default AuthContext;
