+++ frontend/src/auth/context/AuthProvider.tsx
@@ -1,0 +1,24 @@
+import React, {createContext, ReactNode, useState} from "react";
+
+interface AuthProviderProps {
+  children: ReactNode;
+}
+
+interface AuthContextType {
+  auth: any;
+  setAuth: React.Dispatch<React.SetStateAction<any>>;
+}
+
+const AuthContext = createContext<AuthContextType | undefined>(undefined);
+
+export function AuthProvider({children}: AuthProviderProps) {
+  const [auth, setAuth] = useState({});
+
+  return (
+    <AuthContext.Provider value={{auth, setAuth}}>
+      {children}
+    </AuthContext.Provider>
+  );
+}
+
+export default AuthContext;
