+++ frontend/src/auth/hooks/useAuth.js
@@ -1,5 +1,5 @@
-import { useContext } from "react";
-import AuthContext from "../context/AuthProvider.jsx";
+import {useContext} from "react";
+import AuthContext from "../context/AuthProvider.tsx";
 
 function useAuth() {
   return useContext(AuthContext);
