+++ frontend/src/public/pages/ErrorPage.tsx
