+++ frontend/vite.config.ts
@@ -1,0 +1,29 @@
+import {defineConfig, loadEnv} from 'vite'
+import react from '@vitejs/plugin-react'
+
+// https://vitejs.dev/config/
+
+export default defineConfig(({mode}) => {
+  const env = loadEnv(mode, process.cwd(), "");
+
+  return {
+    server: { /* only applies to the DEVELOPMENT vite server */
+      https: {
+        key:"../ssl/fake-ssl-key",
+        cert:"../ssl/fake-ssl-cert"
+      },
+      proxy: {
+        "/api": {
+          target: "http://localhost:8080",
+          changeOrigin: true
+        },
+        "/oauth2": {
+          target: "http://localhost:8080",
+          changeOrigin: true
+        }
+      },
+      port: 4430
+    },
+    plugins: [react()]
+  };
+});
